move grafana panel to a nested folder
