+++ grafana-plugin/panel/LICENSE
