+++ grafana-plugin/panel/docker-compose.yml
