+++ grafana-plugin/panel/docs/assets/dashboard.jpg
