+++ grafana-plugin/panel/docs/assets/endpoint.jpg
