+++ grafana-plugin/panel/docs/assets/pluginslist.jpg
