+++ grafana-plugin/panel/docs/assets/settings.jpg
