+++ grafana-plugin/panel/grafana-provisioning/dashboards/dashboard.json
