+++ grafana-plugin/panel/grafana-provisioning/dashboards/main.yml
@@ -6,4 +6,3 @@ providers:
     updateIntervalSeconds: 5
     options:
       path: /etc/grafana/provisioning/dashboards
-
