+++ grafana-plugin/panel/grafana-provisioning/datasources/datasources.yml
