+++ grafana-plugin/panel/grafana.ini
