+++ grafana-plugin/panel/src/SimplePanel.module.css
