+++ grafana-plugin/panel/src/types.ts
